Testing
even more project cleaning

#Issue: strange behavior of the function that reads the Question file
