+++ Exam/Projekt_nr_3.depend
@@ -94,19 +94,21 @@
 	"UI.h"
 	"main.h"
 
-1512769526 source:f:\@github\uph\project_nr_3\exam\src\stack.cpp
+1512821740 source:f:\@github\uph\project_nr_3\exam\src\stack.cpp
 	"Stack.h"
 
-1512769355 f:\@github\uph\project_nr_3\exam\include\stack.h
+1512821569 f:\@github\uph\project_nr_3\exam\include\stack.h
 	<iostream>
+	<ctime>
+	<stdlib.h>
 
 1512767980 source:f:\@github\uph\project_nr_3\exam\src\ui.cpp
 	"UI.h"
 	"main.h"
 
-1512508821 f:\@github\uph\project_nr_3\exam\include\ui.h
+1512772979 f:\@github\uph\project_nr_3\exam\include\ui.h
 
-1512772918 f:\@github\uph\project_nr_3\exam\include\main.h
+1512821553 f:\@github\uph\project_nr_3\exam\include\main.h
 	<string>
 	<ctime>
 	<stdlib.h>
@@ -114,6 +116,7 @@
 	"utilities.h"
 	"UI.h"
 	"Stack.h"
+	"Question.h"
 
 1512492329 f:\@github\uph\project_nr_3\exam\include\utilities.h
 	"main.h"
@@ -123,6 +126,16 @@
 	<stdlib.h>
 	<cstring>
 
-1512771977 source:f:\@github\uph\project_nr_3\exam\src\utilities.cpp
+1512775151 source:f:\@github\uph\project_nr_3\exam\src\utilities.cpp
 	"utilities.h"
+
+1512821520 f:\@github\uph\project_nr_3\exam\include\question.h
+	<iostream>
+	<fstream>
+
+1512822234 source:f:\@github\uph\project_nr_3\exam\src\main.cpp
+	"main.h"
+
+1512821520 source:f:\@github\uph\project_nr_3\exam\src\question.cpp
+	"Question.h"
