+++ Exam/src/Stack.cpp
@@ -25,3 +25,37 @@ void clearStack(NumberStack **root)
         delete temp;
     }
 }
+
+NumberStack* generateUniqueNumber(int n)
+{
+
+    NumberStack *root = NULL, *tempW=NULL;
+
+    int i = 0;
+    int r;
+    bool isUnique = true;
+    while(i < n)
+    {
+        r = ( rand() % 50 ) + 1;
+        tempW = root; // save current stack pointer;
+        while(root) // until root = NULL
+        {
+            int tmp = popStack(&root);
+            if(r == tmp){
+                isUnique = false;
+                break;
+            }
+        }
+        root = tempW; // return to top stack pointer
+
+        if(isUnique){
+            pushStack(&root, r);
+            i++;
+        }else{
+            isUnique = true;
+        }
+    }
+
+   // clearStack(&root);
+    return root;
+}
